# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Final Project: Tiketku",
    "description": "Platform ini merupakan tempat beli dan booking tiket penerbangan secara online, baik one way (sekali penerbangan) dan round trip (pulang pergi). Platform ini membuka dan menyediakan berbagai jenis penerbangan domestik dan mancanegara. Flow / alur platform dibebaskan berdasarkan kreasi tim final project. Terdapat 1 (satu) peran User yaitu login sebagai buyer. Buyer dapat melakukan pencarian, pemilihan, dan melihat histori pemesanan dan detail tiket.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "brahmastabagus@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://be-tiketku-production.up.railway.app/api/v1",
      "description": "Production server"
    },
    {
      "url": "https://743b-103-162-237-17.ngrok-free.app/api/v1/",
      "description": "Stagging server"
    },
    {
      "url": "http://localhost:8080/api/v1/",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/flight/{id}": {
      "get": {
        "tags": [
          "Flight"
        ],
        "summary": "Get flight",
        "description": "Get flight by id.",
        "operationId": "getFlightById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of flight to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get flights by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flights"
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoData"
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Flight"
        ],
        "summary": "Update flight",
        "description": "Update flight data.",
        "operationId": "updateFlight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of flight to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated flight object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flight"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update flight",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Flight"
        ],
        "summary": "Delete flight",
        "description": "Delete flight by id.",
        "operationId": "deleteFlightById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of flight to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete flight",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
    "/airport/{id}": {
      "get": {
        "tags": [
          "Airport"
        ],
        "summary": "Get airport",
        "description": "Get airport by id.",
        "operationId": "getAirportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of airport to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get airports by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airports"
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoData"
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Airport"
        ],
        "summary": "Update airport",
        "description": "Update airport data.",
        "operationId": "updateAirport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of airport to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated airport object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Airport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update airport",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Airport"
        ],
        "summary": "Delete airport",
        "description": "Delete airport by id.",
        "operationId": "deleteAirportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of airport to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete airport",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Flight": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "example": "https://upload.wikimedia.org/wikipedia/commons/b/b0/AdamAir_logo.png"
          },
          "flight_number": {
            "type": "string",
            "example": "1783728"
          },
          "airline": {
            "type": "string",
            "example": "Aerolineas Argentinas"
          },
          "source_airport": {
            "type": "integer",
            "example": 1
          },
          "destination_airport": {
            "type": "integer",
            "example": 2
          },
          "departure_airport": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "arrival_date": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "capacity": {
            "type": "integer",
            "example": 50
          },
          "economy_class_price": {
            "type": "integer",
            "example": 50000
          },
          "business_class_price": {
            "type": "integer",
            "example": 50000
          },
          "first_class_price": {
            "type": "integer",
            "example": 50000
          },
          "premium_price": {
            "type": "integer",
            "example": 50000
          },
          "adult_price_percentage": {
            "type": "integer",
            "example": 100
          },
          "child_price_percentage": {
            "type": "integer",
            "example": 60
          },
          "baby_price_percentage": {
            "type": "integer",
            "example": 0
          }
        },
        "required": [
          "image",
          "flight_number",
          "airline",
          "source_airport",
          "destination_airport",
          "departure_airport",
          "arrival_date",
          "capacity",
          "economy_class_price",
          "business_class_price",
          "first_class_price",
          "premium_price",
          "adult_price_percentage",
          "child_price_percentage",
          "baby_price_percentage"
        ]
      },
      "Flights": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "image": {
            "type": "string",
            "example": "https://upload.wikimedia.org/wikipedia/commons/b/b0/AdamAir_logo.png"
          },
          "flight_number": {
            "type": "string",
            "example": "1783728"
          },
          "airline": {
            "type": "string",
            "example": "Aerolineas Argentinas"
          },
          "source_airport": {
            "type": "integer",
            "example": 1
          },
          "destination_airport": {
            "type": "integer",
            "example": 2
          },
          "departure_airport": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "arrival_date": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "capacity": {
            "type": "integer",
            "example": 50
          },
          "economy_class_price": {
            "type": "integer",
            "example": 50000
          },
          "business_class_price": {
            "type": "integer",
            "example": 50000
          },
          "first_class_price": {
            "type": "integer",
            "example": 50000
          },
          "premium_price": {
            "type": "integer",
            "example": 50000
          },
          "adult_price_percentage": {
            "type": "integer",
            "example": 100
          },
          "child_price_percentage": {
            "type": "integer",
            "example": 60
          },
          "baby_price_percentage": {
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "source": {
            "$ref": "#/components/schemas/Airports"
          },
          "destination": {
            "$ref": "#/components/schemas/Airports"
          }
        }
      },
      "Airport": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Istanbul Atatürk Airport"
          },
          "code": {
            "type": "string",
            "example": "IST"
          },
          "city": {
            "type": "string",
            "example": "Istanbul"
          },
          "country": {
            "type": "string",
            "example": "Turkey"
          }
        },
        "required": [
          "name",
          "code",
          "city",
          "country"
        ]
      },
      "Airports": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "name": {
            "type": "string",
            "example": "Istanbul Atatürk Airport"
          },
          "code": {
            "type": "string",
            "example": "IST"
          },
          "city": {
            "type": "string",
            "example": "Istanbul"
          },
          "country": {
            "type": "string",
            "example": "Turkey"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          }
        }
      },
...
      "NoData": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "data": {
            "type": "array",
            "example": [ ]
          }
        }
      },
...
      "Res404": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "URL tidak ditemukan"
          }
        }
      },
      "Res400": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "\"name\" is required"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}